requirements.txt
langchain>=1.0
langchain-openai>=0.2
langgraph>=1.0
deepagents>=0.4.0
python-dotenv>=1.0
tavily>=1.0
pydantic>=2.0
fastapi>=0.115
uvicorn[standard]>=0.32
